requirements.txt
torch>=1.7.0
transformers>=4.5.0
pandas>=1.2.0
numpy>=1.19.0
scikit-learn>=0.24.0
python-dotenv>=0.19.0
tqdm>=4.50.0
flask>=2.0.0
requests>=2.25.0